Added Assimp to content processor
Removed large unnecessary PDBs

Fix post build step

Added a rendering frontend
Added a serialization system
Tested both

Syntax fixes

Made the serialization streams interfaces. Made the most sense in terms of use case.

Added a primary test for reading and writing a compiled mesh asset
Fixed a bug in the buddy allocator where we were using the wrong size to determine what memory pages to commit.

Added extension support to the content processor.
Added an isDirectory to the platform library.

Added path command lines to the content processor

Loaded the compiled mesh into the graphics sample

Changed the serialization API to remove the inheritance. It was unnecessary. If we need to make it polymorphic in the future, we can do it then.
Improved the loading code in the graphics sample to use our platform API for file mapping.